requirements.txt
huggingface_hub
requests
python-dotenv
pytest
